(ref): Renamed Component to Drawables
